requirements.txt
fastapi
uvicorn
supabase
python-dotenv
langchain-ollama
langchain-core
requests